feat(BotServer): change the behaviour when there is no peers in the webChannel
Add BotServer.removeWebChannel() method. If a peer wants to invite a bot to the webChannel with an

id (for example id=1) and the Bot already has the instance of the webChannel with such id, then he

checks if there are no peers in this webChannel then he accepts the invitation, otherwise he ignores

it.

BREAKING CHANGE: change BotServer.listen(options) into BotServer.start(), BotServer.stopListen()

into BotServer.stop(). The server options should be passed to the BotServer constructor instead of

BotServer.listen method.