refactor: improve architecture, make easier to maintain and debug
The refactoring  is not finished yet. The main goal of this is to allow create a WebSocket over
Signaling server in order to allow to bot to connect to Signaling server. Simplify, improve and
complete the algorithm of connection creation (selection between WebSocket and RTCDataChannel) in
ChannelBuilder. Add a notion of 'stream'. Two stream exist WebChannel and Signaling streams.